:Update the main page component
