+++ app/javascript/components/Layout.js
@@ -1,0 +1,10 @@
+import { Outlet } from 'react-router-dom';
+
+const Layout = () => (
+  <>
+    <h1>Layout page</h1>
+    <Outlet />
+  </>
+);
+
+export default Layout;
