+++ app/javascript/redux/flights/flightsSlice.js
@@ -3,17 +3,16 @@ import axios from 'axios';
 
 const url = 'http://localhost:3000/api/v1/flights'
 
-// const flights = [
-//     {name: 'Brazil', image: 'https://images.pexels.com/photos/16129715/pexels-photo-16129715.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300'},
-//     {name: 'Egypt', image: 'https://images.pexels.com/photos/3958516/pexels-photo-3958516.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300'},
-//     {name: 'America', image: 'https://images.pexels.com/photos/356844/pexels-photo-356844.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300'},
-//     {name: 'Morocco', image: 'https://images.pexels.com/photos/6945915/pexels-photo-6945915.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300'},
-//     {name: 'Cameroon', image: 'https://images.pexels.com/photos/17290979/pexels-photo-17290979.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300'},
-//     {name: 'Nigeria', image: 'https://images.pexels.com/photos/16237519/pexels-photo-16237519.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300'},
-//     {name: 'Mexico', image: 'https://images.pexels.com/photos/17806066/pexels-photo-17806066.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300'},
-//     {name: 'Ghana', image: 'https://images.pexels.com/photos/6567674/pexels-photo-6567674.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300'},
-//   ]
-
+const flights = [
+    {name: 'Brazil', picture: 'https://images.pexels.com/photos/16129715/pexels-photo-16129715.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300', base_price: 1500, available_slots: 10, reserved: false},
+    {name: 'Egypt', picture: 'https://images.pexels.com/photos/3958516/pexels-photo-3958516.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300', base_price: 1100, available_slots: 10, reserved: false},
+    {name: 'America', picture: 'https://images.pexels.com/photos/356844/pexels-photo-356844.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300', base_price: 2000, available_slots: 10, reserved: false},
+    {name: 'Morocco', picture: 'https://images.pexels.com/photos/6945915/pexels-photo-6945915.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300', base_price: 1200, available_slots: 10, reserved: false},
+    {name: 'Cameroon', picture: 'https://images.pexels.com/photos/17290979/pexels-photo-17290979.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300', base_price: 1400, available_slots: 10, reserved: false},
+    {name: 'Nigeria', picture: 'https://images.pexels.com/photos/16237519/pexels-photo-16237519.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300', base_price: 1300, available_slots: 10, reserved: false},
+    {name: 'Mexico', picture: 'https://images.pexels.com/photos/17806066/pexels-photo-17806066.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300', base_price: 1050, available_slots: 10, reserved: false},
+    {name: 'Ghana', picture: 'https://images.pexels.com/photos/6567674/pexels-photo-6567674.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=200&w=300', base_price: 1000, available_slots: 10, reserved: false},
+  ]
 export const fetchFlights = createAsyncThunk('flights/fetchFlights', async () => {
     try {
         const response = await axios.get(url);
@@ -24,11 +23,11 @@ export const fetchFlights = createAsyncThunk('flights/fetchFlights', async () =>
 });
 
 const initialState = {
-    flights: [],
+    flights: [...flights],
     status: 'idle',
     error: null,
     cwidth: 0,
-    active: null
+    active: ''
 };
 
 const flightsSlice = createSlice({
